use full private key in export link
